fix(factory_reset_tools): remove flutter library dependencies for cli
