+++ packages/factory_reset_tools/pubspec.yaml
@@ -20,6 +20,7 @@ dependencies:
   flutter_svg: ^2.0.10
   handy_window: ^0.4.0
   intl: ^0.18.1
+  meta: ^1.11.0
   retry: ^3.1.2
   ubuntu_localizations: ^0.3.5
   ubuntu_utils: ^0.1.0
